I added the function headers that, based on my model, are necessary in client.h, and I implemented the Client_connect function in client.c. This function establishes a connection with the server specified in the Client structure. Then, I realized the need to verify that the client was working, so I began developing the server.
The server consists of a simple TCP program that listens for incoming connections on a
specified address and port. When a client connects, the server creates a new thread to
handle that connection independently.
In each connection, the server reads a message from the client and displays it on the console.

Finally, I created the Cargo.toml file to compile the server, as well as the CMakeLists.txt
file to compile the client.